um: time: fix userspace detection during tick accounting
The cpu usage timekeeping triggered by the tick either credits time to
either the kernel or currently running userspace process.
On UML tick interrupts firing while userspace is running are not marked
correctly so this bookkeeping is broken.
Specifically 'user_tick' passed to account_process_tick() is always false.

The full callchain is this:

um_timer()
  -> handle_irq_event()
    -> tick_handle_periodic()
      -> tick_periodic()
        -> update_process_times(user_mode(get_irq_regs()))
          -> account_process_tick(task, false)

The issue can be reproduced easily with the selftest
tools/testing/selftests/timers/posix_timers.c,
which hangs in the ITIMER_VIRTUAL/SIGVTALRM testcase.

Fix up the IRQ regs by correctly setting is_user in the IRQ registers.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
---
I'm not familiar with UML, so this is probably not the right fix.
Feel free to treat it as bugreport.