timens: Remove hard dependency on vDSO
Time namespaces require special support in the vDSO, as it is implemented
in the generic vDSO library. However the dependency on the vDSO library
prevents configurations without any vDSO from enabling time namespaces.

As SPARC64 is the only architecture which still implements their own vDSO,
only prevent TIME_NS for it specifically, allowing all other architectures
to enable it.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>