s390/time: Set up vDSO datapage later
Upcoming changes to the generic vDSO library will mean that the vDSO
datapage will not yet be usable during time_early_init().

Move the initialization to time_init() which is called later. This is
valid as the value of tod_clock_base.tod does not change during the
lifetime of the machine and the vDSO datapage is only used much later.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>