vdso/datastore: Allocate data pages dynamically
Allocating the datapages as part of the kernel image does not work on
SPARC.  It also prevents switching the data VMA to 'struct page', as the
mm core does not allow those.

Allocate the data pages through the page allocator instead.
These pages are used by both the timekeeping and the random pool
initialization code, so introduce a new toplevel initialization step.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>