vdso/datastore: Allocate data pages dynamically
Allocating the data pages as part of the kernel image does not work on
SPARC. The MMU will through a fault when userspace tries to access them.

Allocate the data pages through the page allocator instead.
Unused pages in the vDSO VMA are still allocated to keep the virtual
addresses aligned. Switch the mapping from PFNs to 'struct page' as that
is required for dynamically allocated pages.
This also aligns the allocation of the datapages with the code
pages and is a prerequisite for mlockall() support.

VM_MIXEDMAP is necessary for the call to vmf_insert_page() in the timens
prefault path to work.

The data pages pages need to be order-0, non-compound pages so that the
mapping to userspace and the different orderings work.

These pages are also used by the timekeeping, random pool and
architecture initialization code. Some of these are running before the
page allocator is available. To keep these subsytems working without
changes, introduce early, statically data storage which will then
replaced by the real one as soon as that is available.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>